Use debian/micro-revision.i.
If we have a debian/micro-revision.i, replace the one in src/or with our
copy so that this will be the revision that ends up in the binary.  This
is an informational only version string, but it'd be kinda nice if it
was (more) accurate nonetheless.  (Backported from 0.2.2.2-alpha-1 from
September 2009.)